Add KSum data structure for efficient sum calculation
Introduces the KSum struct to Data Structures/KSum.cpp, enabling O(log N) addition and removal for maintaining the sum of k smallest elements. Also includes minor updates to Arrays/prefix_sums.cpp and DP/subsequence-pattern/lis.cpp, and moves harmonic series logic to Maths/Series/harmonic series.cpp.

+++ Arrays/prefix_sums.cpp
@@ -24,7 +24,7 @@ struct prefix_sum_2D{
 
 //  1d prefix sum to get sum of a sub array
 struct prefix_sum_1D {
-    vector<int> ps;//using the definition ps[i]= sum of all elements with index<i
+    vector<ll> ps;//using the definition ps[i]= sum of all elements with index<i
     prefix_sum_1D(vector<int>& arr) {
         int n=arr.size();
         ps.assign(n+1,0ll);
